+++ packages/theia-integration/src/browser/channel-connection.ts
@@ -14,10 +14,10 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { Channel, Disposable, DisposableCollection, Emitter, MessageProvider } from '@theia/core';
-import { createMessageConnection, Logger, Message, MessageConnection } from 'vscode-jsonrpc';
-import { AbstractMessageReader, DataCallback, MessageReader } from 'vscode-jsonrpc/lib/messageReader';
-
-import { AbstractMessageWriter, MessageWriter } from 'vscode-jsonrpc/lib/messageWriter';
+import {
+    AbstractMessageReader, AbstractMessageWriter, createMessageConnection, DataCallback, Logger, Message,
+    MessageConnection, MessageReader, MessageWriter
+} from 'vscode-jsonrpc';
 
 // Temporary fix/workaround to enable comparability with Theia >=1.27 until https://github.com/eclipse-theia/theia/issues/11405 is resolved
 
@@ -46,8 +46,8 @@ export class ChannelMessageReader extends AbstractMessageReader implements Messa
         this.toDispose.dispose();
     }
 
-    listen(callback: DataCallback): void {
-        this.onMessageEmitter.event(callback);
+    listen(callback: DataCallback): Disposable {
+        return this.onMessageEmitter.event(callback);
     }
 }
 
@@ -62,10 +62,14 @@ export class ChannelMessageWriter extends AbstractMessageWriter implements Messa
         this.toDispose = channel.onClose(() => this.fireClose());
     }
 
-    write(msg: Message): void {
+    async write(msg: Message): Promise<void> {
         const writeBuffer = this.channel.getWriteBuffer();
         writeBuffer.writeBytes(Buffer.from(JSON.stringify(msg, undefined, 0)));
         writeBuffer.commit();
+    }
+
+    end(): void {
+        this.dispose();
     }
 
     override dispose(): void {